* Exclude signup route from API token interceptor to prevent multiple authorization tokens from being sent

* Hide progress mask on sign up failure

* Update README.md (#66)

* fix manifest XML syntax error (#38)

* Upgrade to gradle 8.9 and java 21 (#68)

* Fix nullability to avoid null pointer exception

* Code Submitted Covers Mage #1326, Mage #1252, and Mage #1356

* Add Comment for Clarity, revert min version

* Fix Error Created during Merge conflict resolution

* add tools-versions file for java open jdk

* Add Scrollview to allow Profile to Scroll when it no longer fits on the screen.

* [mage-1357] - add missing type adapters to instance of gson being used for geoJson serialization to fix issue with inability to sync lines or polygons to mage server

* Bugfix/mage #1363: errors warning typos (#75)

* down from 43 to 12 errors

* down from 12 to 2 errors

* down from 2 to 0 errors

* undid mistake, fixed a lot more warnings

* duplicate

* undoing attempted fix

* undoing another attempted fix

* undoing AndroidID mistake

* Fixing Conflicts

* Fix Rebase Errors

* Fix When Statement

* Update Error Message

* Feature/#1391: Contact Us button (#77)

* Added *Contact Us* button within main menu settings under *About* button.

* added correct 'TO' email address

* Feature/mage 1398 (#78)

* [mage-1398]
Update github actions upload-artifact to v4
Set JVM target to 17 to resolve build workflow failure
Change workflows to not set version code. Update release workflow to set the correct mime type for apks.

* Update version code and name for 7.2.4 release

---------

+++ mage/src/main/java/mil/nga/giat/mage/data/repository/user/UserRepository.kt
@@ -14,17 +14,17 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import mil.nga.giat.mage.R
 import mil.nga.giat.mage.data.datasource.permission.RoleLocalDataSource
+import mil.nga.giat.mage.data.datasource.user.UserLocalDataSource
+import mil.nga.giat.mage.database.dao.MageSqliteOpenHelper
+import mil.nga.giat.mage.database.model.event.Event
+import mil.nga.giat.mage.database.model.user.User
 import mil.nga.giat.mage.di.TokenProvider
 import mil.nga.giat.mage.login.AuthenticationStatus
 import mil.nga.giat.mage.login.AuthorizationStatus
 import mil.nga.giat.mage.network.device.DeviceService
 import mil.nga.giat.mage.network.user.UserService
-import mil.nga.giat.mage.sdk.Compatibility.Companion.isCompatibleWith
-import mil.nga.giat.mage.database.model.event.Event
-import mil.nga.giat.mage.database.model.user.User
-import mil.nga.giat.mage.data.datasource.user.UserLocalDataSource
-import mil.nga.giat.mage.database.dao.MageSqliteOpenHelper
 import mil.nga.giat.mage.network.user.UserWithRoleTypeAdapter
+import mil.nga.giat.mage.sdk.Compatibility.Companion.isCompatibleWith
 import mil.nga.giat.mage.sdk.preferences.PreferenceHelper
 import mil.nga.giat.mage.sdk.utils.DeviceUuidFactory
 import mil.nga.giat.mage.sdk.utils.ISO8601DateFormatFactory
@@ -308,15 +308,20 @@ class UserRepository @Inject constructor(
    private suspend fun syncIcon(user: User) {
       val path = "${MediaUtility.getUserIconDirectory(application)}/${user.id}.png"
       val file = File(path)
-      if (file.exists()) {
+      /// At the moment, I can't see any code that regularly updates User.lastModified
+      /// Aside from the fact that we happen to pull users in their entirety every time we
+      /// do something... That should be fixed and then this code will be absolutely correct.
+      /// Until then, this code is achieving the result of not re-pulling an icon.
+      if (file.exists() && user.fetchedDate.before(user.lastModified)) {
          file.delete()
-      }
-
-      val response = userService.getIcon(user.remoteId)
-      val body = response.body()
-      if (response.isSuccessful && body != null) {
-         saveIcon(body, file)
-         compressIcon(file)
+         val response = userService.getIcon(user.remoteId)
+         val body = response.body()
+         if (response.isSuccessful && body != null) {
+            saveIcon(body, file)
+            compressIcon(file)
+            userLocalDataSource.setIconPath(user, path)
+         }
+      } else {
          userLocalDataSource.setIconPath(user, path)
       }
    }
